add alipay fun and sign

+++ app/src/main/java/hy/zc/wfj/data/OrderListObject.java
@@ -205,7 +205,7 @@ public class OrderListObject implements Serializable{
             return mobilePhone;
         }
 
-        public float getOrdersId() {
+        public int getOrdersId() {
             return ordersId;
         }
